#3 Move api package under pkg
