+++ pkg/api/ecosystem/componentRestClient.go
@@ -3,7 +3,7 @@ package ecosystem
 import (
 	"context"
 	"fmt"
-	v1 "github.com/cloudogu/k8s-component-operator/api/v1"
+	v1 "github.com/cloudogu/k8s-component-operator/pkg/api/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/watch"
