+++ pkg/api/ecosystem/componentRestClient_test.go
@@ -3,7 +3,7 @@ package ecosystem
 import (
 	"context"
 	"encoding/json"
-	v1 "github.com/cloudogu/k8s-component-operator/api/v1"
+	v1 "github.com/cloudogu/k8s-component-operator/pkg/api/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"io"
