+++ pkg/controllers/componentInstallManager_test.go
@@ -3,7 +3,7 @@ package controllers
 import (
 	"context"
 	"fmt"
-	v1 "github.com/cloudogu/k8s-component-operator/api/v1"
+	v1 "github.com/cloudogu/k8s-component-operator/pkg/api/v1"
 	"github.com/cloudogu/k8s-component-operator/pkg/config"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
