+++ pkg/controllers/component_manager_test.go
@@ -2,7 +2,7 @@ package controllers
 
 import (
 	"context"
-	v1 "github.com/cloudogu/k8s-component-operator/api/v1"
+	v1 "github.com/cloudogu/k8s-component-operator/pkg/api/v1"
 	"github.com/cloudogu/k8s-component-operator/pkg/config"
 	"github.com/stretchr/testify/require"
 	"testing"
